mark changed to double

+++ AP_PROJECT/AP_PROJECT/Class/TermCourseStudent.cs
@@ -3,7 +3,7 @@ namespace AP_PROJECT.Class
 {
     public class TermCourseStudent
     {
-        public int Mark { set; get; }
+        public double Mark { set; get; }
         public String ObjectionToMark  { get; set; }
         public String AnswerToObjection { get; set; }
         public TermCourse TermCourse { set; get; }
